requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
distlib==0.3.9
filelock==3.16.1
gibberish-detector==0.1.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
identify==2.6.3
idna==3.10
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
PyPDF2==3.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
ruff==0.8.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.28.0
